Move convert module to subpackage
- Move convert.py module to convert/*.py subpackage
- Split convert.py module into converter.py and cameras/*.py
- Rename *Camera classes to *
