+++ src/glimpse/convert/__init__.py
@@ -1,0 +1,5 @@
+"""Convert between external camera models and the glimpse camera model."""
+from . import cameras
+from .converter import Converter
+
+__all__ = ["cameras", "Converter"]
