+++ src/glimpse/convert/cameras/__init__.py
@@ -1,0 +1,7 @@
+"""External camera models."""
+from .agisoft import Agisoft
+from .matlab import Matlab
+from .opencv import OpenCV
+from .photomodeler import PhotoModeler
+
+__all__ = ["Agisoft", "Matlab", "OpenCV", "PhotoModeler"]
